requirements.txt
requests
browser_cookie3
BeautifulSoup4
colorama
bs4
argparse
psycopg2=2.8.6
python-dotenv==0.15.0
